Extracted buildEarlyExitMatch to shared utility
