+++ packages/url-utils/lib/utils/html-absolute-to-transform-ready.js
@@ -1,30 +1,6 @@
 const htmlTransform = require('./html-transform');
 const absoluteToTransformReady = require('./absolute-to-transform-ready');
-
-function escapeRegExp(string) {
-    return string.replace(/[.*+?^${}()|[\]\\]/g, '\\$&');
-}
-
-function buildEarlyExitMatch(siteUrl, options) {
-    const candidates = [siteUrl, options.imageBaseUrl, options.filesBaseUrl, options.mediaBaseUrl]
-        .filter(Boolean)
-        .map((value) => {
-            let normalized = options.ignoreProtocol ? value.replace(/http:|https:/, '') : value;
-            return normalized.replace(/\/$/, '');
-        })
-        .filter(Boolean)
-        .map(escapeRegExp);
-
-    if (!candidates.length) {
-        return null;
-    }
-
-    if (candidates.length === 1) {
-        return candidates[0];
-    }
-
-    return `(?:${candidates.join('|')})`;
-}
+const {buildEarlyExitMatch} = require('./build-early-exit-match');
 
 const htmlAbsoluteToTransformReady = function (html = '', siteUrl, _options) {
     const defaultOptions = {assetsOnly: false, ignoreProtocol: true};
